Widen KPI/XP economy to long; fix career completion and completed-card overlay
Bug 1: currentExperience/currentKpi/kpiPerClick/kpiPerSecond were int and overflowed past ~2.1B (XP at CEO went negative). Widened the runtime economy and its public API/results to long across GameManager, OfflineProgressResult, StatsTracker, FloatingKpi(Spawner), WorkZoneClick, UIController, CareerTrackView. Multiplier math now rounds via (long)Math.Round((double)...). ScriptableObject fields stay int (max price 1.12B fits). Added a load-time clamp to repair already-overflowed (negative) saves.

Bug 3: career-completed screen never appeared because negative XP failed the currentExperience >= requiredKpi check. Resolved by the long fix.

Bug 2: CompletedOverlay (PLAN DONE) was managed by the card's VerticalLayoutGroup, which stacked it and pushed the bottom-anchored MAX text off the card. Added LayoutElement(ignoreLayout=true) and stretched the overlay to fill the card. Applied to CardButton1.prefab and all three scene instances.

+++ Assets/Scripts/Core/FloatingKpi.cs
@@ -60,12 +60,12 @@ public class FloatingKpi : MonoBehaviour
         }
     }
 
-    public void Init(int amount)
+    public void Init(long amount)
     {
         Init(amount, false);
     }
 
-    public void Init(int amount, bool isCritical)
+    public void Init(long amount, bool isCritical)
     {
         this.isCritical = isCritical;
